Saturday, December 2, 2023 at 8:11 PM West Africa Standard Time
