+++ app/components/landingPage/Explore_comp.js
@@ -1,0 +1,7 @@
+"use client";
+
+function Explore() {
+  return <div className="bg-explore-bg text-white">hello dfhdfdfdhfkh</div>;
+}
+
+export default Explore;
