+++ app/page.js
@@ -1,13 +1,15 @@
 import { FormProvider } from "../context/form/userProfile/FormContext";
 import Form from "./components/form/userProfile/template";
-import Hero from "./components/landingPage/Hero";
+import Hero from "./components/landingPage/Hero_comp";
+import Explore from "./components/landingPage/Explore_comp";
 import Navbar from "./components/landingPage/navbar";
 import PersonaliseScreen from "./components/ui/personaliseScreen/template";
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen items-center bg-grayDark justify-center">
+    <main className=" min-h-screen ">
      <Hero/>
+     <Explore/>
       {/* <PersonaliseScreen
         title="Get Personalized Outfit Recommendations"
         description={`At Mestyle, we're dedicated to helping you look and feel your best. To provide you with the most relevant outfit suggestions, please share some information with us. Rest assured that your data is kept secure and used only to enhance your style journey.`}
